Emacs: evil-easymotion and evil-snipe added
Evil-easymotion is a package that integrates avy like motion with
evil-mode. The key prefix suggested in the github README is SPC.

Evil-snipe replaces the s and S keybinds in normal mode with a line
incremental search. This functionality is generally faster than the
usual f, F, t and T vim keybinds.