Emacs: stoped using ensure-system-package keyword
I've decided to use scripts to install the binary dependencies that
emacs needs instead of the ensure-system-package keyword in
use-package.

I've also removed the exec-path-from-shell-arguments setting in macOS
and linux since it was causing problems with executable-find function
and apparently, starting the shell interactively doesn't slow startup
as it used to do.