Emacs: simplified font configuration
Switched to using set-frame-font to set the default font. This
function changes the default font on every frame and on future
ones. There's also another function to configure faces which is
set-face-attribute but I think the first one is better for changing
the default font. Set-face-attribute can be used to change the
configuration of other fonts as well as the default one.