Dotfiles: change in repo structure
I'm in the process of switching to using GNU Stow to manage
symlinks. Because of that, I've switched the script that created
symlinks for another one that does the same using stow. What prevously
was in src/os/ has been moved to setup/ folder and everything else
that was in src/ has been moved to stow/. The installation scripts
have been adjusted accordingly to ensure everything still works as
intended.

Finally, I've removed github actions, .gitattributes and .editorconfig
since I didn't really use any of those.